restructure namespace
