+++ .editorconfig
@@ -1,0 +1,4 @@
+[*]
+vsspell_section_id = 483b6d021a094f108a18ed12cbc3b634
+vsspell_ignored_words_483b6d021a094f108a18ed12cbc3b634 = File:IgnoredWords.dic
+
