+++ Driver/Driver.csproj
@@ -8,7 +8,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Connect4\Connect4.csproj" />
+    <ProjectReference Include="..\Connect4Backend\Connect4Backend.csproj" />
   </ItemGroup>
 
 </Project>
