+++ UI/Connect4/v1/Board.cs
@@ -1,4 +1,4 @@
-﻿namespace UI
+﻿namespace UI.Connect4.v1
 {
 	public partial class Board : UserControl
 	{
