+++ UI/Connect4/v1/TokenSpace.cs
@@ -9,7 +9,7 @@ namespace UI
 		public int ColumnIndex { get; init; }
 		public int RowIndex { get; init; }
 		private ImageAttributes Attributes { get; set; } = new();
-		private Image Image { get; set; } = new Bitmap(Resources.test);
+		private Image Image { get; set; } = new Bitmap(Resources.Connect4_Token);
 
 		public TokenSpace(int row, int column, Color startColor)
 		{
